Fix Using Static IPs With NAT (#268)
When a static IP is configured and the network mode is set to NAT the
IP is set in the network_interface.addresses field of the domain which
adds it to the libvirt DHCP server. In that case the netplan config in
the guest OS should be set to dynamic to use the DHCP address.

Changing this means that DHCP can be disabled in the static netplan
config which was causing duplicate IP addresses to be assigned.

Co-authored-by: Seumas Dunlop <seumas.dunlop@coralbay.tv>